2da Clase

+++ src/main/java/com/ucoltis/elimas/veterinaria/HistorialClinico.java
@@ -6,10 +6,10 @@ import java.util.List;
 
 public class HistorialClinico {
   private Mascota mascota;
-  private final Integer id;
+  private final Long id;
   private final List<DetalleClinico> detalleClinicos; //getter y final se inicializa en contructor
 
-  public HistorialClinico(Integer id) {
+  public HistorialClinico(Long id) {
     this.id = id;
     detalleClinicos= new ArrayList<>();
   }
@@ -18,9 +18,10 @@ public class HistorialClinico {
     return detalleClinicos;
   }
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
+
   
   public void agregarDetalle(DetalleClinico detalleClinico){
     detalleClinicos.add(detalleClinico);
